refactor(lb): added tests for latency aware, least connections, round robin and sticky cid
