+++ crates/lb/tests/algorithms/round_robin.rs
@@ -15,3 +15,30 @@ fn round_robin_cycles() {
     let picks: Vec<usize> = (0..6).filter_map(|_| rr.pick(&pool)).collect();
     assert_eq!(picks, vec![0, 1, 2, 0, 1, 2]);
 }
+
+#[test]
+fn unhealthy_backends_are_skipped() {
+    let mut pool = BackendPool::new_from_states(vec![
+        create_backend_state("10.0.0.1:1", 1),
+        create_backend_state("10.0.0.2:1", 1),
+    ]);
+
+    pool.mark_failure(0);
+    pool.mark_failure(0);
+    pool.mark_failure(0);
+
+    let mut rr = RoundRobin::new();
+    let pick = rr.pick(&pool).unwrap();
+    assert_eq!(pick, 1);
+}
+
+#[test]
+fn no_healthy_backends_returns_none() {
+    let mut pool = BackendPool::new_from_states(vec![create_backend_state("10.0.0.1:1", 1)]);
+    pool.mark_failure(0);
+    pool.mark_failure(0);
+    pool.mark_failure(0);
+
+    let mut rr = RoundRobin::new();
+    assert!(rr.pick(&pool).is_none());
+}
